\name{nbEset}
\alias{nbEset}
\docType{data}
\title{ Reduced Neuroblastoma data set}
\description{
	Neuroblastoma Data set - reduced in size to comply with BioC package guidelines
}
\usage{data(nbEset)}

\details{
This dataset was retrieved from GEO \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3960} and consists of five reference samples, and ten primary neuroblastoma tumors. Four of the five reference samples \code{GSM2827}, \code{GSM2842}, \code{GSM2883}, and \code{GSM2895} came from a separate dataset, \url{http://www.ncbi.nlm.nih.gov/geo/gds/gds_browse.cgi?gds=181}
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3960}
}

\examples{
data(nbEset)
nbEset
}
\keyword{datasets}
