\name{VAIgsc}
\alias{VAIgsc}
\docType{data}
\title{ Molecular Concepts (Gene Sets) prepared at VAI}
\description{
  A few gene sets compiled at VAI. We have found useful in our analysis.
}
\usage{data(VAIgsc)}
\format{
  The format is:
 chr "VAIgsc"
}

\source{
  Various sources... See individual objects for PMID, GEO accession, etc..
}

\examples{
data(VAIgsc)
summary(VAIgsc)
details(VAIgsc[[1]])

}
\keyword{datasets}
