\name{VAImcs}
\alias{VAImcs}
\docType{data}
\title{ Molecular Concepts prepared at VAI}
\description{
  A few gene sets compiled at VAI. We have found useful in our analysis.
}
\usage{data(VAImcs)}
\format{
 a list of "smc" objects
}
\source{
 See individual concepts for PMID or other source information.
}
\examples{
data(VAImcs)
str(VAImcs)
}
\keyword{datasets}
